tests/wpt/web-platform-tests/dom/nodes/Document-createCDATASection.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>document.createCDATASection must throw in HTML documents</title>
<link rel="help" href="https://dom.spec.whatwg.org/#dom-document-createcdatasection">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

setup({ single_test: true });

assert_throws_dom("NotSupportedError", () => document.createCDATASection("foo"));

done();
</script>
